Removed SystemInfo and modified Images
