+++ Documentation/python/tdw_utils.md
@@ -219,7 +219,7 @@ _Returns:_  A list of commands: The first creates the exterior walls, and the se
 
 ***
 
-#### `save_images(images: Images, filename: str, output_directory="dist", resize_to=None, append_pass: bool = True, width: int = 256, height: int = 256) -> None`
+#### `save_images(images: Images, filename: str, output_directory="dist", resize_to=None, append_pass: bool = True) -> None`
 
 _This is a static function._
 
@@ -234,8 +234,6 @@ The images object includes the pass and extension information.
 | filename | The filename of each image, minus the extension. The image pass will be appended as a prefix. |
 | resize_to | Specify a (width, height) tuple to resize the images to. This is slower than saving as-is. |
 | append_pass | If false, the image pass will _not_ be appended to the filename as a prefix, e.g.: `"0000"`: -> "`0000.jpg"` |
-| width | The expected width in pixels of the image. This is only relevant to the `_depth` and `_depth_simple` passes. |
-| height | The expected height in pixels of the image. This is only relevant to the `_depth` and `_depth_simple` passes. |
 
 ***
 
@@ -307,7 +305,7 @@ _Returns:_ A list of commands to set ALL visual materials on an object to a sing
 
 ***
 
-#### `get_depth_values(image: np.array, depth_pass: str = "_depth", width: int = 256, height: int = 256, uv_starts_on_top: bool = True) -> np.array`
+#### `get_depth_values(image: np.array, depth_pass: str = "_depth", width: int = 256, height: int = 256, uv_starts_at_top: bool = True) -> np.array`
 
 _This is a static function._
 
@@ -319,9 +317,9 @@ The far plane is hardcoded as 100. The near plane is hardcoded as 0.1.
 | --- | --- |
 | image | The image pass as a numpy array. |
 | depth_pass | The type of depth pass. This determines how the values are decoded. Options: `"_depth"`, `"_depth_simple"`. |
-| width | The width of the screen in pixels. |
-| height | The height of the screen in pixels. |
-| uv_starts_on_top | If True, UV coordinates start at the top of the image. See the command `send_system_info`. |
+| width | The width of the screen in pixels. See `Images.get_width()` in the output data documentation. |
+| height | The height of the screen in pixels. See `Images.get_height()` in the output data documentation. |
+| uv_starts_on_top | If True, UV coordinates start at the top of the image. See `Images.get_uv_starts_at_top()` in the output data documentation. |
 
 _Returns:_ An array of depth values.
