+++ Python/example_controllers/depth_shader.py
@@ -2,7 +2,7 @@ from pathlib import Path
 import matplotlib.pyplot as plt
 from tdw.controller import Controller
 from tdw.tdw_utils import TDWUtils
-from tdw.output_data import OutputData, Images, SystemInfo, CameraMatrices
+from tdw.output_data import OutputData, Images, CameraMatrices
 
 
 """
@@ -32,12 +32,10 @@ class DepthShader(Controller):
                          {"$type": "set_pass_masks",
                           "pass_masks": ["_img", depth_pass]},
                          {"$type": "send_images"},
-                         {"$type": "send_system_info"},
                          {"$type": "send_camera_matrices"}])
         resp = self.communicate(commands)
 
         depth_image = None
-        system_info = None
         camera_matrix = None
         images = None
         for i in range(len(resp) - 1):
@@ -48,23 +46,16 @@ class DepthShader(Controller):
                 for j in range(images.get_num_passes()):
                     if images.get_pass_mask(j) == "_depth":
                         depth_image = images.get_image(j)
-            # Get the system info.
-            elif r_id == "sysi":
-                system_info = SystemInfo(resp[i])
             # Get the camera matrix.
             elif r_id == "cama":
                 camera_matrix = CameraMatrices(resp[i]).get_camera_matrix()
         # Save the image. Note that because we're using depth images, the width and height must be defined.
-        # (The width and height default to 256x256, which is TDW's default screen size.)
-        width = system_info.get_width()
-        height = system_info.get_height()
-        TDWUtils.save_images(images=images, output_directory="D:/depth_shader", filename="0", append_pass=True,
-                             width=width, height=height)
+        TDWUtils.save_images(images=images, output_directory="D:/depth_shader", filename="0", append_pass=True)
         # Get the depth values of each pixel.
         depth = TDWUtils.get_depth_values(image=depth_image,
-                                          width=width,
-                                          height=height,
-                                          uv_starts_on_top=system_info.get_uv_starts_at_top())
+                                          width=images.get_width(),
+                                          height=images.get_height(),
+                                          uv_starts_at_top=images.get_uv_starts_at_top())
         print(depth)
         # Get a point cloud and write it to disk.
         point_cloud_filename = "point_cloud.txt"
