--- Python/tdw/FBOutput/SystemInfo.py
@@ -1,70 +1,0 @@
-# automatically generated by the FlatBuffers compiler, do not modify
-
-# namespace: FBOutput
-
-import tdw.flatbuffers
-
-class SystemInfo(object):
-    __slots__ = ['_tab']
-
-    @classmethod
-    def GetRootAsSystemInfo(cls, buf, offset):
-        n = tdw.flatbuffers.encode.Get(tdw.flatbuffers.packer.uoffset, buf, offset)
-        x = SystemInfo()
-        x.Init(buf, n + offset)
-        return x
-
-    # SystemInfo
-    def Init(self, buf, pos):
-        self._tab = tdw.flatbuffers.table.Table(buf, pos)
-
-    # SystemInfo
-    def UvStartsAtTop(self):
-        o = tdw.flatbuffers.number_types.UOffsetTFlags.py_type(self._tab.Offset(4))
-        if o != 0:
-            return bool(self._tab.Get(tdw.flatbuffers.number_types.BoolFlags, o + self._tab.Pos))
-        return False
-
-    # SystemInfo
-    def SupportsAudio(self):
-        o = tdw.flatbuffers.number_types.UOffsetTFlags.py_type(self._tab.Offset(6))
-        if o != 0:
-            return bool(self._tab.Get(tdw.flatbuffers.number_types.BoolFlags, o + self._tab.Pos))
-        return False
-
-    # SystemInfo
-    def Width(self):
-        o = tdw.flatbuffers.number_types.UOffsetTFlags.py_type(self._tab.Offset(8))
-        if o != 0:
-            return self._tab.Get(tdw.flatbuffers.number_types.Int32Flags, o + self._tab.Pos)
-        return 0
-
-    # SystemInfo
-    def Height(self):
-        o = tdw.flatbuffers.number_types.UOffsetTFlags.py_type(self._tab.Offset(10))
-        if o != 0:
-            return self._tab.Get(tdw.flatbuffers.number_types.Int32Flags, o + self._tab.Pos)
-        return 0
-
-    # SystemInfo
-    def Gpu(self):
-        o = tdw.flatbuffers.number_types.UOffsetTFlags.py_type(self._tab.Offset(12))
-        if o != 0:
-            return self._tab.String(o + self._tab.Pos)
-        return None
-
-    # SystemInfo
-    def Os(self):
-        o = tdw.flatbuffers.number_types.UOffsetTFlags.py_type(self._tab.Offset(14))
-        if o != 0:
-            return self._tab.String(o + self._tab.Pos)
-        return None
-
-def SystemInfoStart(builder): builder.StartObject(6)
-def SystemInfoAddUvStartsAtTop(builder, uvStartsAtTop): builder.PrependBoolSlot(0, uvStartsAtTop, 0)
-def SystemInfoAddSupportsAudio(builder, supportsAudio): builder.PrependBoolSlot(1, supportsAudio, 0)
-def SystemInfoAddWidth(builder, width): builder.PrependInt32Slot(2, width, 0)
-def SystemInfoAddHeight(builder, height): builder.PrependInt32Slot(3, height, 0)
-def SystemInfoAddGpu(builder, gpu): builder.PrependUOffsetTRelativeSlot(4, tdw.flatbuffers.number_types.UOffsetTFlags.py_type(gpu), 0)
-def SystemInfoAddOs(builder, os): builder.PrependUOffsetTRelativeSlot(5, tdw.flatbuffers.number_types.UOffsetTFlags.py_type(os), 0)
-def SystemInfoEnd(builder): return builder.EndObject()
